[STYLE]: 검색 결과 기본 UI 추가
