+++ src/components/search/Search.tsx
@@ -2,7 +2,7 @@ import { useRef, useState } from "react";
 import styled from "@emotion/styled";
 
 import { Input } from "~/components/shared/Input";
-import { DefaultRecommend } from "~/components/search";
+import { SearchingInfo } from "./SearchingInfo";
 
 const Wrapper = styled.div`
   position: relative;
@@ -12,9 +12,10 @@ const Styled = {
   Wrapper,
 };
 
-const SearchBar = () => {
-  const ref = useRef<HTMLInputElement | null>(null);
+const Search = () => {
+  const inputRef = useRef<HTMLInputElement | null>(null);
   const [isFocus, setIsFocus] = useState(false);
+  const [inputValue, setInputValue] = useState("");
 
   const handleFocus = () => {
     setIsFocus(true);
@@ -25,7 +26,7 @@ const SearchBar = () => {
   };
 
   const handleEnterPress = () => {
-    const searchData = ref.current?.value.trim();
+    const searchData = inputRef.current?.value.trim();
 
     if (!searchData) {
       return alert("검색어를 입력해주세요.");
@@ -33,22 +34,24 @@ const SearchBar = () => {
   };
 
   const handleChange = () => {
+    setInputValue(inputRef.current?.value.trim() || "");
     // 실시간으로 검색 결과 UI에 검색어를 렌더링 해야 함
   };
 
   const handleClearInput = () => {
-    const { current } = ref;
+    const { current } = inputRef;
     if (current === null) {
       return;
     }
 
     current.value = "";
+    setInputValue("");
   };
 
   return (
     <Styled.Wrapper>
       <Input
-        ref={ref}
+        ref={inputRef}
         hasSearchIcon
         isFocus={isFocus}
         placeholder="질환명을 입력해 주세요."
@@ -59,9 +62,9 @@ const SearchBar = () => {
         onEnterPress={handleEnterPress}
         style={{ width: "490px" }}
       />
-      {isFocus && <DefaultRecommend />}
+      {isFocus && <SearchingInfo inputValue={inputValue} />}
     </Styled.Wrapper>
   );
 };
 
-export default SearchBar;
+export default Search;
