+++ src/components/search/index.ts
@@ -1,2 +1,1 @@
-export { default as SearchBar } from "./SearchBar";
-export { default as DefaultRecommend } from "./DefaultRecommend";
+export { default as Search } from "./Search";
